requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
xlrd==2.0.1
openpyxl
python-multipart
passlib
bcrypt